+++ pkg/log/pattern.go
@@ -4,16 +4,11 @@ import (
 	"bytes"
 	"fmt"
 	"io"
+	"path"
 	"strings"
 	"sync"
 	"time"
 )
-
-// Render render log output
-type Render interface {
-	Render(io.Writer, map[string]interface{}) error
-	RenderString(map[string]interface{}) string
-}
 
 var patternMap = map[string]func(map[string]interface{}) string{
 	"T": longTime,
@@ -25,8 +20,8 @@ var patternMap = map[string]func(map[string]interface{}) string{
 	"i": keyFactory(_instanceID),
 	"e": keyFactory(_deplyEnv),
 	"z": keyFactory(_zone),
-	"S": keyFactory(_source),
-	"s": keyFactory(_source),
+	"S": longSource,
+	"s": shortSource,
 	"M": message,
 }
 
@@ -78,6 +73,7 @@ func (p *pattern) Render(w io.Writer, d map[string]interface{}) error {
 	for _, f := range p.funcs {
 		buf.WriteString(f(d))
 	}
+
 	_, err := buf.WriteTo(w)
 	return err
 }
@@ -112,6 +108,20 @@ func keyFactory(key string) func(map[string]interface{}) string {
 		}
 		return ""
 	}
+}
+
+func longSource(d map[string]interface{}) string {
+	if fn, ok := d[_source].(string); ok {
+		return fn
+	}
+	return "unknown:0"
+}
+
+func shortSource(d map[string]interface{}) string {
+	if fn, ok := d[_source].(string); ok {
+		return path.Base(fn)
+	}
+	return "unknown:0"
 }
 
 func longTime(map[string]interface{}) string {