+++ pkg/net/rpc/warden/client.go
@@ -10,6 +10,9 @@ import (
 	"sync"
 	"time"
 
+	"github.com/bilibili/kratos/pkg/net/rpc/warden/resolver"
+	"github.com/bilibili/kratos/pkg/net/rpc/warden/resolver/direct"
+
 	"github.com/bilibili/kratos/pkg/conf/env"
 	"github.com/bilibili/kratos/pkg/conf/flagvar"
 	"github.com/bilibili/kratos/pkg/ecode"
@@ -18,8 +21,6 @@ import (
 	"github.com/bilibili/kratos/pkg/net/netutil/breaker"
 	"github.com/bilibili/kratos/pkg/net/rpc/warden/balancer/p2c"
 	"github.com/bilibili/kratos/pkg/net/rpc/warden/internal/status"
-	"github.com/bilibili/kratos/pkg/net/rpc/warden/resolver"
-	"github.com/bilibili/kratos/pkg/net/rpc/warden/resolver/direct"
 	"github.com/bilibili/kratos/pkg/net/trace"
 	xtime "github.com/bilibili/kratos/pkg/time"
 
@@ -36,9 +37,11 @@ var _grpcTarget flagvar.StringVars
 var (
 	_once           sync.Once
 	_defaultCliConf = &ClientConfig{
-		Dial:    xtime.Duration(time.Second * 10),
-		Timeout: xtime.Duration(time.Millisecond * 250),
-		Subset:  50,
+		Dial:              xtime.Duration(time.Second * 10),
+		Timeout:           xtime.Duration(time.Millisecond * 250),
+		Subset:            50,
+		KeepAliveInterval: xtime.Duration(time.Second * 60),
+		KeepAliveTimeout:  xtime.Duration(time.Second * 20),
 	}
 	_defaultClient *Client
 )
@@ -51,20 +54,24 @@ func baseMetadata() metadata.MD {
 	return gmd
 }
 
+// Register direct resolver by default to handle direct:// scheme.
 func init() {
 	resolver.Register(direct.New())
 }
 
 // ClientConfig is rpc client conf.
 type ClientConfig struct {
-	Dial     xtime.Duration
-	Timeout  xtime.Duration
-	Breaker  *breaker.Config
-	Method   map[string]*ClientConfig
-	Clusters []string
-	Zone     string
-	Subset   int
-	NonBlock bool
+	Dial                   xtime.Duration
+	Timeout                xtime.Duration
+	Breaker                *breaker.Config
+	Method                 map[string]*ClientConfig
+	Clusters               []string
+	Zone                   string
+	Subset                 int
+	NonBlock               bool
+	KeepAliveInterval      xtime.Duration
+	KeepAliveTimeout       xtime.Duration
+	KeepAliveWithoutStream bool
 }
 
 // Client is the framework's client side instance, it contains the ctx, opt and interceptors.
@@ -74,8 +81,19 @@ type Client struct {
 	breaker *breaker.Group
 	mutex   sync.RWMutex
 
-	opt      []grpc.DialOption
+	opts     []grpc.DialOption
 	handlers []grpc.UnaryClientInterceptor
+}
+
+// TimeoutCallOption timeout option.
+type TimeoutCallOption struct {
+	*grpc.EmptyCallOption
+	Timeout time.Duration
+}
+
+// WithTimeoutCallOption can override the timeout in ctx and the timeout in the configuration file
+func WithTimeoutCallOption(timeout time.Duration) *TimeoutCallOption {
+	return &TimeoutCallOption{&grpc.EmptyCallOption{}, timeout}
 }
 
 // handle returns a new unary client interceptor for OpenTracing\Logging\LinkTimeout.
@@ -107,11 +125,24 @@ func (c *Client) handle() grpc.UnaryClientInterceptor {
 		c.mutex.RUnlock()
 		brk := c.breaker.Get(method)
 		if err = brk.Allow(); err != nil {
-			statsClient.Incr(method, "breaker")
+			_metricClientReqCodeTotal.Inc(method, "breaker")
 			return
 		}
 		defer onBreaker(brk, &err)
-		_, ctx, cancel = conf.Timeout.Shrink(ctx)
+		var timeOpt *TimeoutCallOption
+		for _, opt := range opts {
+			var tok bool
+			timeOpt, tok = opt.(*TimeoutCallOption)
+			if tok {
+				break
+			}
+		}
+		if timeOpt != nil && timeOpt.Timeout > 0 {
+			ctx, cancel = context.WithTimeout(nmd.WithContext(ctx), timeOpt.Timeout)
+		} else {
+			_, ctx, cancel = conf.Timeout.Shrink(ctx)
+		}
+
 		defer cancel()
 		nmd.Range(ctx,
 			func(key string, value interface{}) {
@@ -144,9 +175,10 @@ func (c *Client) handle() grpc.UnaryClientInterceptor {
 
 func onBreaker(breaker breaker.Breaker, err *error) {
 	if err != nil && *err != nil {
-		if ecode.ServerErr.Equal(*err) || ecode.ServiceUnavailable.Equal(*err) || ecode.Deadline.Equal(*err) || ecode.LimitExceed.Equal(*err) {
+		if ecode.EqualError(ecode.ServerErr, *err) || ecode.EqualError(ecode.ServiceUnavailable, *err) || ecode.EqualError(ecode.Deadline, *err) || ecode.EqualError(ecode.LimitExceed, *err) {
 			breaker.MarkFailed()
 			return
+
 		}
 	}
 	breaker.MarkSuccess()
@@ -166,7 +198,6 @@ func NewClient(conf *ClientConfig, opt ...grpc.DialOption) *Client {
 	}
 	c.UseOpt(grpc.WithBalancerName(p2c.Name))
 	c.UseOpt(opt...)
-	c.Use(c.recovery(), clientLogging(), c.handle())
 	return c
 }
 
@@ -192,6 +223,12 @@ func (c *Client) SetConfig(conf *ClientConfig) (err error) {
 	}
 	if conf.Subset <= 0 {
 		conf.Subset = 50
+	}
+	if conf.KeepAliveInterval <= 0 {
+		conf.KeepAliveInterval = xtime.Duration(time.Second * 60)
+	}
+	if conf.KeepAliveTimeout <= 0 {
+		conf.KeepAliveTimeout = xtime.Duration(time.Second * 20)
 	}
 
 	// FIXME(maojian) check Method dial/timeout
@@ -221,21 +258,33 @@ func (c *Client) Use(handlers ...grpc.UnaryClientInterceptor) *Client {
 }
 
 // UseOpt attachs a global grpc DialOption to the Client.
-func (c *Client) UseOpt(opt ...grpc.DialOption) *Client {
-	c.opt = append(c.opt, opt...)
+func (c *Client) UseOpt(opts ...grpc.DialOption) *Client {
+	c.opts = append(c.opts, opts...)
 	return c
 }
 
-// Dial creates a client connection to the given target.
-// Target format is scheme://authority/endpoint?query_arg=value
-// example: direct://default/192.168.1.1:9000,192.168.1.2:9001
-func (c *Client) Dial(ctx context.Context, target string, opt ...grpc.DialOption) (conn *grpc.ClientConn, err error) {
+func (c *Client) cloneOpts() []grpc.DialOption {
+	dialOptions := make([]grpc.DialOption, len(c.opts))
+	copy(dialOptions, c.opts)
+	return dialOptions
+}
+
+func (c *Client) dial(ctx context.Context, target string, opts ...grpc.DialOption) (conn *grpc.ClientConn, err error) {
+	dialOptions := c.cloneOpts()
 	if !c.conf.NonBlock {
-		c.opt = append(c.opt, grpc.WithBlock())
+		dialOptions = append(dialOptions, grpc.WithBlock())
 	}
-	c.opt = append(c.opt, grpc.WithInsecure())
-	c.opt = append(c.opt, grpc.WithUnaryInterceptor(c.chainUnaryClient()))
-	c.opt = append(c.opt, opt...)
+	dialOptions = append(dialOptions, opts...)
+
+	// init default handler
+	var handlers []grpc.UnaryClientInterceptor
+	handlers = append(handlers, c.recovery())
+	handlers = append(handlers, clientLogging(dialOptions...))
+	handlers = append(handlers, c.handlers...)
+	// NOTE: c.handle must be a last interceptor.
+	handlers = append(handlers, c.handle())
+
+	dialOptions = append(dialOptions, grpc.WithUnaryInterceptor(chainUnaryClient(handlers)))
 	c.mutex.RLock()
 	conf := c.conf
 	c.mutex.RUnlock()
@@ -268,43 +317,39 @@ func (c *Client) Dial(ctx context.Context, target string, opt ...grpc.DialOption
 		}
 		target = u.String()
 	}
-	if conn, err = grpc.DialContext(ctx, target, c.opt...); err != nil {
+	if conn, err = grpc.DialContext(ctx, target, dialOptions...); err != nil {
 		fmt.Fprintf(os.Stderr, "warden client: dial %s error %v!", target, err)
 	}
 	err = errors.WithStack(err)
 	return
 }
 
+// Dial creates a client connection to the given target.
+// Target format is scheme://authority/endpoint?query_arg=value
+// example: discovery://default/account.account.service?cluster=shfy01&cluster=shfy02
+func (c *Client) Dial(ctx context.Context, target string, opts ...grpc.DialOption) (conn *grpc.ClientConn, err error) {
+	opts = append(opts, grpc.WithInsecure())
+	return c.dial(ctx, target, opts...)
+}
+
 // DialTLS creates a client connection over tls transport to the given target.
-func (c *Client) DialTLS(ctx context.Context, target string, file string, name string) (conn *grpc.ClientConn, err error) {
+func (c *Client) DialTLS(ctx context.Context, target string, file string, name string, opts ...grpc.DialOption) (conn *grpc.ClientConn, err error) {
 	var creds credentials.TransportCredentials
 	creds, err = credentials.NewClientTLSFromFile(file, name)
 	if err != nil {
 		err = errors.WithStack(err)
 		return
 	}
-	c.opt = append(c.opt, grpc.WithBlock())
-	c.opt = append(c.opt, grpc.WithTransportCredentials(creds))
-	c.opt = append(c.opt, grpc.WithUnaryInterceptor(c.chainUnaryClient()))
-	c.mutex.RLock()
-	conf := c.conf
-	c.mutex.RUnlock()
-	if conf.Dial > 0 {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, time.Duration(conf.Dial))
-		defer cancel()
-	}
-	conn, err = grpc.DialContext(ctx, target, c.opt...)
-	err = errors.WithStack(err)
-	return
+	opts = append(opts, grpc.WithTransportCredentials(creds))
+	return c.dial(ctx, target, opts...)
 }
 
 // chainUnaryClient creates a single interceptor out of a chain of many interceptors.
 //
 // Execution is done in left-to-right order, including passing of context.
 // For example ChainUnaryClient(one, two, three) will execute one before two before three.
-func (c *Client) chainUnaryClient() grpc.UnaryClientInterceptor {
-	n := len(c.handlers)
+func chainUnaryClient(handlers []grpc.UnaryClientInterceptor) grpc.UnaryClientInterceptor {
+	n := len(handlers)
 	if n == 0 {
 		return func(ctx context.Context, method string, req, reply interface{},
 			cc *grpc.ClientConn, invoker grpc.UnaryInvoker, opts ...grpc.CallOption) error {
@@ -323,9 +368,9 @@ func (c *Client) chainUnaryClient() grpc.UnaryClientInterceptor {
 				return invoker(ictx, imethod, ireq, ireply, ic, iopts...)
 			}
 			i++
-			return c.handlers[i](ictx, imethod, ireq, ireply, ic, chainHandler, iopts...)
+			return handlers[i](ictx, imethod, ireq, ireply, ic, chainHandler, iopts...)
 		}
 
-		return c.handlers[0](ctx, method, req, reply, cc, chainHandler, opts...)
+		return handlers[0](ctx, method, req, reply, cc, chainHandler, opts...)
 	}
 }