+++ pkg/ratelimit/bbr/bbr.go
@@ -17,10 +17,11 @@ import (
 
 var (
 	cpu         int64
-	decay       = 0.75
+	decay       = 0.95
+	initTime    = time.Now()
 	defaultConf = &Config{
-		Window:       time.Second * 5,
-		WinBucket:    50,
+		Window:       time.Second * 10,
+		WinBucket:    100,
 		CPUThreshold: 800,
 	}
 )
@@ -31,14 +32,17 @@ func init() {
 	go cpuproc()
 }
 
+// cpu = cpuᵗ⁻¹ * decay + cpuᵗ * (1 - decay)	
 func cpuproc() {
+	ticker := time.NewTicker(time.Millisecond * 250)
 	defer func() {
+		ticker.Stop()
 		if err := recover(); err != nil {
 			log.Error("rate.limit.cpuproc() err(%+v)", err)
 			go cpuproc()
 		}
 	}()
-	ticker := time.NewTicker(time.Millisecond * 250)
+
 	// EMA algorithm: https://blog.csdn.net/m0_38106113/article/details/81542863
 	for range ticker.C {
 		stat := &cpustat.Stat{}
@@ -64,11 +68,14 @@ type Stat struct {
 type BBR struct {
 	cpu             cpuGetter
 	passStat        metric.RollingCounter
-	rtStat          metric.RollingGauge
+	rtStat          metric.RollingCounter
 	inFlight        int64
 	winBucketPerSec int64
 	conf            *Config
-	prevDrop        time.Time
+	prevDrop        atomic.Value
+	prevDropHit     int32
+	rawMaxPASS      int64
+	rawMinRt        int64
 }
 
 // Config contains configs of bbr limiter.
@@ -82,9 +89,13 @@ type Config struct {
 }
 
 func (l *BBR) maxPASS() int64 {
-	val := int64(l.passStat.Reduce(func(iterator metric.Iterator) float64 {
+	rawMaxPass := atomic.LoadInt64(&l.rawMaxPASS)
+	if rawMaxPass > 0 && l.passStat.Timespan() < 1 {
+		return rawMaxPass
+	}
+	rawMaxPass = int64(l.passStat.Reduce(func(iterator metric.Iterator) float64 {
 		var result = 1.0
-		for iterator.Next() {
+		for i := 1; iterator.Next() && i < l.conf.WinBucket; i++ {
 			bucket := iterator.Bucket()
 			count := 0.0
 			for _, p := range bucket.Points {
@@ -94,16 +105,21 @@ func (l *BBR) maxPASS() int64 {
 		}
 		return result
 	}))
-	if val == 0 {
-		return 1
+	if rawMaxPass == 0 {
+		rawMaxPass = 1
 	}
-	return val
+	atomic.StoreInt64(&l.rawMaxPASS, rawMaxPass)
+	return rawMaxPass
 }
 
 func (l *BBR) minRT() int64 {
-	val := l.rtStat.Reduce(func(iterator metric.Iterator) float64 {
+	rawMinRT := atomic.LoadInt64(&l.rawMinRt)
+	if rawMinRT > 0 && l.rtStat.Timespan() < 1 {
+		return rawMinRT
+	}
+	rawMinRT = int64(math.Ceil(l.rtStat.Reduce(func(iterator metric.Iterator) float64 {
 		var result = math.MaxFloat64
-		for iterator.Next() {
+		for i := 1; iterator.Next() && i < l.conf.WinBucket; i++ {
 			bucket := iterator.Bucket()
 			if len(bucket.Points) == 0 {
 				continue
@@ -116,8 +132,12 @@ func (l *BBR) minRT() int64 {
 			result = math.Min(result, avg)
 		}
 		return result
-	})
-	return int64(math.Ceil(val))
+	})))
+	if rawMinRT <= 0 {
+		rawMinRT = 1
+	}
+	atomic.StoreInt64(&l.rawMinRt, rawMinRT)
+	return rawMinRT
 }
 
 func (l *BBR) maxFlight() int64 {
@@ -125,15 +145,31 @@ func (l *BBR) maxFlight() int64 {
 }
 
 func (l *BBR) shouldDrop() bool {
-	inFlight := atomic.LoadInt64(&l.inFlight)
-	maxInflight := l.maxFlight()
 	if l.cpu() < l.conf.CPUThreshold {
-		if time.Now().Sub(l.prevDrop) <= 1000*time.Millisecond {
-			return inFlight > 1 && inFlight > maxInflight
+		prevDrop, _ := l.prevDrop.Load().(time.Duration)
+		if prevDrop == 0 {
+			return false
 		}
+		if time.Since(initTime)-prevDrop <= time.Second {
+			if atomic.LoadInt32(&l.prevDropHit) == 0 {
+				atomic.StoreInt32(&l.prevDropHit, 1)
+			}
+			inFlight := atomic.LoadInt64(&l.inFlight)
+			return inFlight > 1 && inFlight > l.maxFlight()
+		}
+		l.prevDrop.Store(time.Duration(0))
 		return false
 	}
-	return inFlight > 1 && inFlight > maxInflight
+	inFlight := atomic.LoadInt64(&l.inFlight)
+	drop := inFlight > 1 && inFlight > l.maxFlight()
+	if drop {
+		prevDrop, _ := l.prevDrop.Load().(time.Duration)
+		if prevDrop != 0 {
+			return drop
+		}
+		l.prevDrop.Store(time.Since(initTime))
+	}
+	return drop
 }
 
 // Stat tasks a snapshot of the bbr limiter.
@@ -155,13 +191,12 @@ func (l *BBR) Allow(ctx context.Context, opts ...limit.AllowOption) (func(info l
 		opt.Apply(&allowOpts)
 	}
 	if l.shouldDrop() {
-		l.prevDrop = time.Now()
 		return nil, ecode.LimitExceed
 	}
 	atomic.AddInt64(&l.inFlight, 1)
-	stime := time.Now()
+	stime := time.Since(initTime)
 	return func(do limit.DoneInfo) {
-		rt := int64(time.Since(stime) / time.Millisecond)
+		rt := int64((time.Since(initTime) - stime) / time.Millisecond)
 		l.rtStat.Add(rt)
 		atomic.AddInt64(&l.inFlight, -1)
 		switch do.Op {
@@ -181,7 +216,7 @@ func newLimiter(conf *Config) limit.Limiter {
 	size := conf.WinBucket
 	bucketDuration := conf.Window / time.Duration(conf.WinBucket)
 	passStat := metric.NewRollingCounter(metric.RollingCounterOpts{Size: size, BucketDuration: bucketDuration})
-	rtStat := metric.NewRollingGauge(metric.RollingGaugeOpts{Size: size, BucketDuration: bucketDuration})
+	rtStat := metric.NewRollingCounter(metric.RollingCounterOpts{Size: size, BucketDuration: bucketDuration})
 	cpu := func() int64 {
 		return atomic.LoadInt64(&cpu)
 	}
@@ -191,7 +226,6 @@ func newLimiter(conf *Config) limit.Limiter {
 		passStat:        passStat,
 		rtStat:          rtStat,
 		winBucketPerSec: int64(time.Second) / (int64(conf.Window) / int64(conf.WinBucket)),
-		prevDrop:        time.Unix(0, 0),
 	}
 	return limiter
 }