+++ tool/kratos-protoc/bm.go
@@ -7,8 +7,8 @@ import (
 )
 
 const (
-	_getBMGen = "go get github.com/bilibili/kratos/tool/protobuf/protoc-gen-bm"
-	_bmProtoc = "protoc --proto_path=%s --proto_path=%s --proto_path=%s --bm_out=explicit_http=true:."
+	_getBMGen = "go get -u github.com/bilibili/kratos/tool/protobuf/protoc-gen-bm"
+	_bmProtoc = "protoc --proto_path=%s --proto_path=%s --proto_path=%s --bm_out=:."
 )
 
 func installBMGen() error {