+++ tool/kratos/template.go
@@ -72,8 +72,8 @@ import (
 	"syscall"
 	"time"
 
-	"{{.Name}}/internal/server/http"
-	"{{.Name}}/internal/service"
+	"{{.ModuleName}}/internal/server/http"
+	"{{.ModuleName}}/internal/service"
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 	"github.com/bilibili/kratos/pkg/log"
 )
@@ -122,9 +122,9 @@ import (
 	"syscall"
 	"time"
 
-	"{{.Name}}/internal/server/grpc"
-	"{{.Name}}/internal/server/http"
-	"{{.Name}}/internal/service"
+	"{{.ModuleName}}/internal/server/grpc"
+	"{{.ModuleName}}/internal/server/http"
+	"{{.ModuleName}}/internal/service"
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 	"github.com/bilibili/kratos/pkg/log"
 )
@@ -142,8 +142,7 @@ func main() {
 	httpSrv := http.New(svc)
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, syscall.SIGHUP, syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT)
-	for {
-		s := <-c
+	for s := range c {
 		log.Info("get a signal %s", s.String())
 		switch s {
 		case syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT:
@@ -189,8 +188,14 @@ import (
 	xtime "github.com/bilibili/kratos/pkg/time"
 )
 
-// Dao dao.
-type Dao struct {
+// Dao dao interface
+type Dao interface {
+   Close()
+   Ping(ctx context.Context) (err error)
+}
+
+// dao dao.
+type dao struct {
 	db          *sql.DB
 	redis       *redis.Pool
 	redisExpire int32
@@ -205,7 +210,7 @@ func checkErr(err error) {
 }
 
 // New new a dao and return.
-func New() (dao *Dao) {
+func New() (Dao) {
 	var (
 		dc struct {
 			Demo *sql.Config
@@ -222,7 +227,7 @@ func New() (dao *Dao) {
 	checkErr(paladin.Get("mysql.toml").UnmarshalTOML(&dc))
 	checkErr(paladin.Get("redis.toml").UnmarshalTOML(&rc))
 	checkErr(paladin.Get("memcache.toml").UnmarshalTOML(&mc))
-	dao = &Dao{
+	return &dao{
 		// mysql
 		db: sql.NewMySQL(dc.Demo),
 		// redis
@@ -232,18 +237,17 @@ func New() (dao *Dao) {
 		mc:       memcache.New(mc.Demo),
 		mcExpire: int32(time.Duration(mc.DemoExpire) / time.Second),
 	}
-	return
 }
 
 // Close close the resource.
-func (d *Dao) Close() {
+func (d *dao) Close() {
 	d.mc.Close()
 	d.redis.Close()
 	d.db.Close()
 }
 
 // Ping ping the resource.
-func (d *Dao) Ping(ctx context.Context) (err error) {
+func (d *dao) Ping(ctx context.Context) (err error) {
 	if err = d.pingMC(ctx); err != nil {
 		return
 	}
@@ -253,14 +257,14 @@ func (d *Dao) Ping(ctx context.Context) (err error) {
 	return d.db.Ping(ctx)
 }
 
-func (d *Dao) pingMC(ctx context.Context) (err error) {
+func (d *dao) pingMC(ctx context.Context) (err error) {
 	if err = d.mc.Set(ctx, &memcache.Item{Key: "ping", Value: []byte("pong"), Expiration: 0}); err != nil {
 		log.Error("conn.Set(PING) error(%v)", err)
 	}
 	return
 }
 
-func (d *Dao) pingRedis(ctx context.Context) (err error) {
+func (d *dao) pingRedis(ctx context.Context) (err error) {
 	conn := d.redis.Get(ctx)
 	defer conn.Close()
 	if _, err = conn.Do("SET", "ping", "pong"); err != nil {
@@ -279,14 +283,14 @@ func (d *Dao) pingRedis(ctx context.Context) (err error) {
 import (
 	"context"
 
-	"{{.Name}}/internal/dao"
+	"{{.ModuleName}}/internal/dao"
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 )
 
 // Service service.
 type Service struct {
 	ac  *paladin.Map
-	dao *dao.Dao
+	dao dao.Dao
 }
 
 // New new a service and return.
@@ -319,8 +323,8 @@ import (
 	"context"
 	"fmt"
 
-	pb "{{.Name}}/api"
-	"{{.Name}}/internal/dao"
+	pb "{{.ModuleName}}/api"
+	"{{.ModuleName}}/internal/dao"
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 
 	"github.com/golang/protobuf/ptypes/empty"
@@ -329,7 +333,7 @@ import (
 // Service service.
 type Service struct {
 	ac  *paladin.Map
-	dao *dao.Dao
+	dao dao.Dao
 }
 
 // New new a service and return.
@@ -376,8 +380,8 @@ func (s *Service) Close() {
 import (
 	"net/http"
 
-	"{{.Name}}/internal/model"
-	"{{.Name}}/internal/service"
+	"{{.ModuleName}}/internal/model"
+	"{{.ModuleName}}/internal/service"
 
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 	"github.com/bilibili/kratos/pkg/log"
@@ -437,9 +441,9 @@ func howToStart(c *bm.Context) {
 import (
 	"net/http"
 
-	pb "{{.Name}}/api"
-	"{{.Name}}/internal/model"
-	"{{.Name}}/internal/service"
+	pb "{{.ModuleName}}/api"
+	"{{.ModuleName}}/internal/model"
+	"{{.ModuleName}}/internal/service"
 
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 	"github.com/bilibili/kratos/pkg/log"
@@ -537,39 +541,23 @@ message HelloResp {
 type Kratos struct {
 	Hello string
 }`
-	_tplGoMod = `module {{.Name}}
+	_tplGoMod = `module {{.ModuleName}}
 
 go 1.12
 
 require (
-	github.com/bilibili/kratos v0.0.0-20190417043059-a08e0f771592
+	github.com/bilibili/kratos v0.2.2
 	github.com/gogo/protobuf v1.2.1
-	github.com/golang/protobuf v1.3.1
-	golang.org/x/net v0.0.0-20190420063019-afa5a82059c6
-	google.golang.org/grpc v1.20.1
-)
-
-replace (
-	cloud.google.com/go => github.com/googleapis/google-cloud-go v0.26.0
-	golang.org/x/crypto => github.com/golang/crypto v0.0.0-20190123085648-057139ce5d2b
-	golang.org/x/lint => github.com/golang/lint v0.0.0-20181026193005-c67002cb31c3
-	golang.org/x/net => github.com/golang/net v0.0.0-20190420063019-afa5a82059c6
-	golang.org/x/oauth2 => github.com/golang/oauth2 v0.0.0-20180821212333-d2e6202438be
-	golang.org/x/sync => github.com/golang/sync v0.0.0-20181108010431-42b317875d0f
-	golang.org/x/sys => github.com/golang/sys v0.0.0-20180905080454-ebe1bf3edb33
-	golang.org/x/text => github.com/golang/text v0.3.0
-	golang.org/x/time => github.com/golang/time v0.0.0-20190308202827-9d24e82272b4
-	golang.org/x/tools => github.com/golang/tools v0.0.0-20190328211700-ab21143f2384
-	google.golang.org/appengine => github.com/golang/appengine v1.1.0
-	google.golang.org/genproto => github.com/google/go-genproto v0.0.0-20180817151627-c66870c02cf8
-	google.golang.org/grpc => github.com/grpc/grpc-go v1.20.1
+	github.com/golang/protobuf v1.3.2
+	golang.org/x/net v0.0.0-20190628185345-da137c7871d7
+	google.golang.org/grpc v1.22.0
 )
 `
 	_tplGRPCServer = `package grpc
 
 import (
-	pb "{{.Name}}/api"
-	"{{.Name}}/internal/service"
+	pb "{{.ModuleName}}/api"
+	"{{.ModuleName}}/internal/service"
 	"github.com/bilibili/kratos/pkg/conf/paladin"
 	"github.com/bilibili/kratos/pkg/net/rpc/warden"
 )
@@ -592,5 +580,9 @@ func New(svc *service.Service) *warden.Server {
 	}
 	return ws
 }
+`
+	_tplGogen = `package api
+
+//go:generate kratos tool protoc --swagger --grpc --bm api.proto
 `
 )